# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Piccolo TPV API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables/{tableId}/order:
    get:
      operationId: getTableOrder
      tags: [orders]
      summary: Get active order for an occupied table
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table with its active order and items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableWithOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No active order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{orderId}/items:
    post:
      operationId: addOrderItem
      tags: [orders]
      summary: Add an item to an order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /order-items/{itemId}:
    delete:
      operationId: deleteOrderItem
      tags: [orders]
      summary: Delete a draft order item
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Item deleted
        "400":
          description: Item not in draft status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/send:
    post:
      operationId: sendOrder
      tags: [orders]
      summary: Send order to production (creates kitchen tasks)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: No draft items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products/{productId}/modifiers:
    get:
      operationId: getProductModifiers
      tags: [categories]
      summary: Get modifier groups and options for a product
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Modifier groups for the product
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModifierGroup"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /order-items/{itemId}/details:
    patch:
      operationId: updateOrderItemDetails
      tags: [orders]
      summary: Update notes, allergy info and modifiers on a line item
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderItemDetailsInput"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{orderId}/pase:
    post:
      operationId: markOrderPase
      tags: [kds]
      summary: Mark order as collected/served from pase
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaseActionInput"
      responses:
        "200":
          description: Pase action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: string
        zoneId:
          type: string
        name:
          type: string
        capacity:
          type: integer
        status:
          type: string
        x:
          type: integer
        y:
          type: integer
        shape:
          type: string
      required: [id, zoneId, name, capacity, status, x, y, shape]
    Order:
      type: object
      properties:
        id:
          type: string
        tableId:
          type: string
        employeeId:
          type: string
        status:
          type: string
        createdAt:
          type: string
        sentAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [id, tableId, employeeId, status, createdAt]
    TableWithOrder:
      type: object
      properties:
        table:
          $ref: "#/components/schemas/Table"
        order:
          $ref: "#/components/schemas/Order"
      required: [table, order]
    OrderItem:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        status:
          type: string
        notes:
          type: string
        allergyNote:
          type: string
        hasAllergy:
          type: boolean
        modifiers:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemModifier"
        createdAt:
          type: string
      required: [id, orderId, productId, productName, quantity, unitPrice, status, createdAt, notes, allergyNote, hasAllergy]
    OrderItemModifier:
      type: object
      properties:
        id:
          type: string
        orderItemId:
          type: string
        modifierId:
          type: string
        modifierName:
          type: string
        priceDelta:
          type: string
      required: [id, orderItemId, modifierName, priceDelta]
    ModifierGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        required:
          type: boolean
        maxSelect:
          type: integer
        modifiers:
          type: array
          items:
            $ref: "#/components/schemas/ModifierOption"
      required: [id, name, required, maxSelect, modifiers]
    ModifierOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        priceDelta:
          type: string
      required: [id, name, priceDelta]
    UpdateOrderItemDetailsInput:
      type: object
      properties:
        notes:
          type: string
        allergyNote:
          type: string
        hasAllergy:
          type: boolean
        modifiers:
          type: array
          items:
            type: object
            properties:
              modifierId:
                type: string
              modifierName:
                type: string
              priceDelta:
                type: string
            required: [modifierName, priceDelta]
...
    AddOrderItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        notes:
          type: string
      required: [productId, quantity]
...
    PaseActionInput:
      type: object
      properties:
        action:
          type: string
          enum: [collected, served]
      required: [action]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]