# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Piccolo TPV API
 * OpenAPI spec version: 0.2.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: string;
  tableId: string;
  employeeId: string;
  status: string;
  createdAt: string;
  sentAt?: string;
  items?: OrderItem[];
}